Delete unpractical Formatters
1. Delete KeyIdI.
Reason: KeyIdI can be used only in one scenario, so there is no point in
making it as a struct.
2. Delete OutputFormatters.
Reason: Completely unpractical and one scenario specific. For example:
what if I need to output double with "dd" postfix in one place and with
"D" in another? I can't simple change operator<<, I need to create a new
struct. Because of this the idea of OutputFormatters is bad.